refactor: organise code per feature, infrastructure and shared
