--- composeApp/src/commonMain/kotlin/birds/data/datasource/BirdsRemoteDataSource.kt
@@ -1,7 +1,0 @@
-package birds.data.datasource
-
-import birds.data.model.BirdImage
-
-internal interface BirdsRemoteDataSource {
-    suspend fun getBirdImages(): List<BirdImage>
-}
