--- composeApp/src/commonMain/kotlin/birds/data/model/BirdImage.kt
@@ -1,10 +1,0 @@
-package birds.data.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class BirdImage(
-    val category: String,
-    val path: String,
-    val author: String,
-)
