--- composeApp/src/commonMain/kotlin/birds/data/model/GetBirdsError.kt
@@ -1,6 +1,0 @@
-package birds.data.model
-
-internal sealed class GetBirdsError {
-    data object Network : GetBirdsError()
-    data object Unknown : GetBirdsError()
-}
