--- composeApp/src/commonMain/kotlin/birds/data/repository/BirdsRepository.kt
@@ -1,9 +1,0 @@
-package birds.data.repository
-
-import arrow.core.Either
-import birds.data.model.GetBirdsError
-import birds.ui.model.Bird
-
-internal interface BirdsRepository {
-    suspend fun getBirds(): Either<GetBirdsError, List<Bird>>
-}
