--- composeApp/src/commonMain/kotlin/birds/ui/contract/BirdsUiState.kt
@@ -1,19 +1,0 @@
-package birds.ui.contract
-
-import birds.ui.model.Bird
-
-internal sealed interface BirdsUiState {
-    data object Loading : BirdsUiState
-    data object Failure : BirdsUiState
-    data class Success(
-        val images: List<Bird>,
-        val selectedCategory: String?,
-    ) : BirdsUiState {
-        val categories: Set<String> = images.map { it.category }.toSet()
-        val selectedImages: List<Bird> = images.filter { it.category == selectedCategory }
-    }
-
-    companion object {
-        val initialState = Loading
-    }
-}
