--- composeApp/src/commonMain/kotlin/core/common/Constants.kt
@@ -1,4 +1,0 @@
-package core.common
-
-const val WHITE_SPACE = " "
-const val BY = "by"
