--- composeApp/src/commonMain/kotlin/core/data/datasource/DataException.kt
@@ -1,5 +1,0 @@
-package core.data.datasource
-
-internal sealed class DataException : Exception() {
-    data class Network(val code: Int, override val message: String) : DataException()
-}
