--- composeApp/src/commonMain/kotlin/core/ui/theme/BirdAppTheme.kt
@@ -1,23 +1,0 @@
-package core.ui.theme
-
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.MaterialTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-
-@Composable
-fun AppTheme(
-    content: @Composable () -> Unit,
-) {
-    MaterialTheme(
-        colors = MaterialTheme.colors.copy(primary = Color.Black),
-        shapes = MaterialTheme.shapes.copy(
-            small = RoundedCornerShape(0.dp),
-            medium = RoundedCornerShape(0.dp),
-            large = RoundedCornerShape(0.dp),
-        ),
-    ) {
-        content()
-    }
-}
