--- composeApp/src/commonMain/kotlin/core/ui/theme/Padding.kt
@@ -1,9 +1,0 @@
-package core.ui.theme
-
-import androidx.compose.ui.unit.dp
-
-val extraSmallPadding = 4.dp
-val smallPadding = 8.dp
-val mediumPadding = 16.dp
-val largePadding = 20.dp
-val extraLargePadding = 24.dp
