+++ composeApp/src/commonMain/kotlin/feature/birds/data/datasource/BirdsRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package feature.birds.data.datasource
+
+import feature.birds.data.model.BirdImage
+
+internal interface BirdsRemoteDataSource {
+    suspend fun getBirdImages(): List<BirdImage>
+}
