+++ composeApp/src/commonMain/kotlin/feature/birds/data/model/BirdImage.kt
@@ -1,0 +1,10 @@
+package feature.birds.data.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class BirdImage(
+    val category: String,
+    val path: String,
+    val author: String,
+)
