+++ composeApp/src/commonMain/kotlin/feature/birds/data/model/GetBirdsError.kt
@@ -1,0 +1,6 @@
+package feature.birds.data.model
+
+internal sealed class GetBirdsError {
+    data object Network : GetBirdsError()
+    data object Unknown : GetBirdsError()
+}
