+++ composeApp/src/commonMain/kotlin/infrastructure/core/common/Constants.kt
@@ -1,0 +1,3 @@
+package infrastructure.core.common
+
+const val WHITE_SPACE = " "
