+++ composeApp/src/commonMain/kotlin/infrastructure/core/data/datasource/DataException.kt
@@ -1,0 +1,5 @@
+package infrastructure.core.data.datasource
+
+internal sealed class DataException : Exception() {
+    data class Network(val code: Int, override val message: String) : DataException()
+}
