+++ composeApp/src/commonMain/kotlin/infrastructure/coreui/composable/CoreLoadingDialog.kt
@@ -1,0 +1,51 @@
+package infrastructure.coreui.composable
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.window.Dialog
+import androidx.compose.ui.window.DialogProperties
+
+@Composable
+fun CoreLoadingDialog(
+    isLoading: Boolean,
+    modifier: Modifier = Modifier,
+) {
+    if (isLoading) {
+        Box(modifier = modifier) {
+            Dialog(
+                onDismissRequest = {},
+                properties = DialogProperties(
+                    dismissOnBackPress = false,
+                    dismissOnClickOutside = false,
+                ),
+            ) {
+                Box(
+                    contentAlignment = Alignment.Center,
+                    modifier = Modifier
+                        .size(100.dp)
+                        .background(
+                            color = Color.White,
+                            shape = RoundedCornerShape(0.dp),
+                        ),
+                ) {
+                    CircularProgressIndicator(
+                        modifier = Modifier.size(45.dp),
+                        strokeWidth = 5.dp,
+                    )
+                    CircularProgressIndicator(
+                        modifier = Modifier.size(30.dp),
+                        strokeWidth = 3.dp,
+                    )
+                }
+            }
+        }
+    }
+}
