+++ composeApp/src/commonMain/kotlin/infrastructure/coreui/theme/BirdAppTheme.kt
@@ -1,0 +1,23 @@
+package infrastructure.coreui.theme
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun AppTheme(
+    content: @Composable () -> Unit,
+) {
+    MaterialTheme(
+        colors = MaterialTheme.colors.copy(primary = Color.Black),
+        shapes = MaterialTheme.shapes.copy(
+            small = RoundedCornerShape(0.dp),
+            medium = RoundedCornerShape(0.dp),
+            large = RoundedCornerShape(0.dp),
+        ),
+    ) {
+        content()
+    }
+}
