+++ composeApp/src/commonMain/kotlin/infrastructure/coreui/theme/Padding.kt
@@ -1,0 +1,9 @@
+package infrastructure.coreui.theme
+
+import androidx.compose.ui.unit.dp
+
+val extraSmallPadding = 4.dp
+val smallPadding = 8.dp
+val mediumPadding = 16.dp
+val largePadding = 20.dp
+val extraLargePadding = 24.dp
